Move useProjection to generic hooks, drop sphere deps
